requirements.txt
clamd
flask
gunicorn[gevent]
google-cloud-storage
